// src/prompt/gen/prompt_v1_pb.ts
// @generated by protoc-gen-es v2.16.0 with parameter "target=ts"
// @generated from file prompt_v1.proto (package safety_core.prompt.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file prompt_v1.proto.
 */
export const file_prompt_v1: GenFile = /*@__PURE__*/
  fileDesc("Cg9wcm9tcHRfdjEucHJvdG8SFXNhZmV0eV9jb3JlLnByb21wdC52MSJFChdQZXJtaXNzaW9uUHJvbXB0UmVxdWVzdBIKCgJpZBgBIAEoCRINCgV0aXRsZRgCIAEoCRIPCgdtZXNzYWdlGAMgASgJIjgKGFBlcm1pc3Npb25Qcm9tcHRSZXNwb25zZRIKCgJpZBgBIAEoCRIQCghhcHByb3ZlZBgCIAEoCGIGcHJvdG8z");

/**
 * @generated from message safety_core.prompt.v1.PermissionPromptRequest
 */
export type PermissionPromptRequest = Message<"safety_core.prompt.v1.PermissionPromptRequest"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: string title = 2;
   */
  title: string;

  /**
   * @generated from field: string message = 3;
   */
  message: string;
};

/**
 * Describes the message safety_core.prompt.v1.PermissionPromptRequest.
 * Use `create(PermissionPromptRequestSchema)` to create a new message.
 */
export const PermissionPromptRequestSchema: GenMessage<PermissionPromptRequest> = /*@__PURE__*/
  messageDesc(file_prompt_v1, 0);

/**
 * @generated from message safety_core.prompt.v1.PermissionPromptResponse
 */
export type PermissionPromptResponse = Message<"safety_core.prompt.v1.PermissionPromptResponse"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: bool approved = 2;
   */
  approved: boolean;
};

/**
 * Describes the message safety_core.prompt.v1.PermissionPromptResponse.
 * Use `create(PermissionPromptResponseSchema)` to create a new message.
 */
export const PermissionPromptResponseSchema: GenMessage<PermissionPromptResponse> = /*@__PURE__*/
  messageDesc(file_prompt_v1, 1);